+++ src/cloudscribe.Core.Identity/IProcessAccountLoginRules.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace cloudscribe.Core.Identity
+{
+    public interface IProcessAccountLoginRules
+    {
+        Task ProcessAccountLoginRules(LoginResultTemplate template);
+    }
+}
